Fix ARM build on Mac. Failed on line 3459, Utils::Minimum had incompatible types ...
Review URL: https://codereview.chromium.org//581173002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@40445 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/intermediate_language_arm.cc
@@ -3370,7 +3370,7 @@ void BinaryInt32OpInstr::EmitNativeCode(FlowGraphCompiler* compiler) {
   if (locs()->in(1).IsConstant()) {
     const Object& constant = locs()->in(1).constant();
     ASSERT(constant.IsSmi());
-    const int32_t value = Smi::Cast(constant).Value();
+    const intptr_t value = Smi::Cast(constant).Value();
     switch (op_kind()) {
       case Token::kADD: {
         if (deopt == NULL) {
